corrected the input shaper
the input shaper was usign the individual x and y inputs which would actually change the directiion. now it calculates the magnitude and retains the direction. Also, ctre automatically has a deadband so none is needed from us. Also increased the p gain since the modules were sluggish in the sim